[@foal/cli] Move the scripts to src/.
