+++ packages/cli/src/generate/generators/app/create-app.ts
@@ -64,14 +64,13 @@ export function createApp({ name, sessionSecret }:
   mkdirIfNotExists(`${names.kebabName}/public`);
   copyFileFromTemplates('app/public/logo.png', `${names.kebabName}/public/logo.png`);
 
-  // Scripts
-  mkdirIfNotExists(`${names.kebabName}/scripts`);
-  copyFileFromTemplates('app/scripts/create-users.ts', `${names.kebabName}/scripts/create-users.ts`);
-
   // Src
   mkdirIfNotExists(`${names.kebabName}/src`);
   copyFileFromTemplates('app/src/index.ts', `${names.kebabName}/src/index.ts`);
   copyFileFromTemplates('app/src/test.ts', `${names.kebabName}/src/test.ts`);
+
+  mkdirIfNotExists(`${names.kebabName}/src/scripts`);
+  copyFileFromTemplates('app/src/scripts/create-users.ts', `${names.kebabName}/src/scripts/create-users.ts`);
 
   mkdirIfNotExists(`${names.kebabName}/src/app`);
   copyFileFromTemplates('app/src/app/app.module.ts', `${names.kebabName}/src/app/app.module.ts`);
